+++ src/lib/forms/CheckboxInline.svelte
@@ -1,3 +1,0 @@
-<div class="flex">
-	<slot />
-</div>