+++ src/lib/forms/Input.svelte
@@ -1,30 +1,19 @@
 <script lang="ts">
+	import classNames from 'classnames';
 	import type { InputType } from '../types';
-	import generateId from '../utils/generateId.js';
 	export let type: InputType = 'text';
 	export let value: string = '';
-	export let name: string = '';
-	export let id: string = generateId();
-	export let label: string = '';
-	export let required: boolean = false;
-	export let placeholder: string = '';
-	export let size: 'sm:text-md' | 'text-sm' | 'sm:text-xs' = 'text-sm';
-	let padding: string = '';
-	export let inputClass: string = `bg-gray-50 border border-gray-300 text-gray-900 ${size} rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full ${padding} dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500`;
-	export let labelClass: string = 'block mb-4 text-sm font-medium text-gray-900 dark:text-gray-300';
-	export let disabled: boolean = false;
-	export let readonly: boolean = false;
-	export let helper: string = '';
-	export let helperClass: string = 'text-sm text-gray-500 dark:text-gray-400';
+	export let size: 'sm' | 'md' | 'lg' = 'md';
+	export let inputClass: string =
+		'block w-full border disabled:cursor-not-allowed disabled:opacity-50 rounded-lg';
+	export let color: 'base' | 'green' | 'red' = 'base';
+	const colorClasses = {
+		base: 'bg-gray-50 border-gray-300 text-gray-900 focus:border-blue-500 focus:ring-blue-500 dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder-gray-400 dark:focus:border-blue-500 dark:focus:ring-blue-500',
+		green:
+			'border-green-500 bg-green-50 text-green-900 placeholder-green-700 focus:border-green-500 focus:ring-green-500 dark:border-green-400 dark:bg-green-100 dark:focus:border-green-500 dark:focus:ring-green-500',
+		red: 'border-red-500 bg-red-50 text-red-900 placeholder-red-700 focus:border-red-500 focus:ring-red-500 dark:border-red-400 dark:bg-red-100 dark:focus:border-red-500 dark:focus:ring-red-500'
+	};
 	export let ref: HTMLElement = null;
-
-	if (size === 'sm:text-md') {
-		padding = 'p-4';
-	} else if (size === 'text-sm') {
-		padding = 'p-2.5';
-	} else {
-		padding = 'p-2';
-	}
 
 	// you need to this to avoid 2-way binding
 	function setType(node) {
@@ -32,23 +21,23 @@
 	}
 </script>
 
-<div class={$$props.class ? $$props.class : ''}>
-	{#if label}
-		<label for={id} class={labelClass}>{label}</label>
-	{/if}
-	<input
-		bind:value
-		bind:this={ref}
-		{name}
-		use:setType
-		{id}
-		class={inputClass}
-		{placeholder}
-		{required}
-		{disabled}
-		{readonly}
-	/>
-	{#if helper}
-		<p class={helperClass}>{@html helper}</p>
-	{/if}
-</div>
+<input
+	{...$$restProps}
+	bind:value
+	bind:this={ref}
+	use:setType
+	class={classNames(
+		inputClass,
+		colorClasses[color],
+		{
+			// 'pl-10': icon !== null,
+			// 'rounded-lg': !($$slots.addon || addon),
+			// 'rounded-r-lg': $$slots.addon || addon,
+			// 'shadow-sm dark:shadow-sm-light': shadow,
+			'p-2 sm:text-xs': size === 'sm',
+			'p-2.5 text-sm': size === 'md',
+			'sm:text-md p-4': size === 'lg'
+		},
+		$$props.class
+	)}
+/>