+++ src/lib/index.ts
@@ -59,16 +59,18 @@ export { default as SitemapFooter } from './footer/SitemapFooter.svelte';
 
 // Forms
 export { default as Checkbox } from './forms/Checkbox.svelte';
-export { default as CheckboxInline } from './forms/CheckboxInline.svelte';
 export { default as Fileupload } from './forms/Fileupload.svelte';
 export { default as FloatingLabelInput } from './forms/FloatingLabelInput.svelte';
+export { default as Helper } from './forms/Helper.svelte';
 export { default as Iconinput } from './forms/Iconinput.svelte';
 export { default as Input } from './forms/Input.svelte';
+export { default as Label } from './forms/Label.svelte';
 export { default as RadioInline } from './forms/RadioInline.svelte';
 export { default as Radio } from './forms/Radio.svelte';
 export { default as Range } from './forms/Range.svelte';
 export { default as Search } from './forms/Search.svelte';
 export { default as SimpleSearch } from './forms/SimpleSearch.svelte';
+export { default as VoiceSearch } from './forms/VoiceSearch.svelte';
 export { default as Select } from './forms/Select.svelte';
 export { default as Textarea } from './forms/Textarea.svelte';
 export { default as Toggle } from './forms/Toggle.svelte';